new lib for getExecutedToolTimeline
